refactor: Address PR comments.
