+++ lib/actions/form.js
@@ -19,9 +19,6 @@ const {
   getUrlParams,
   planParamsToQuery
 } = coreUtils.query
-const { getItem, randId } = coreUtils.storage
-const { OTP_API_TIME_FORMAT } = coreUtils.time
-const { isMobile } = coreUtils.ui
 
 export const settingQueryParam = createAction('SET_QUERY_PARAM')
 export const clearActiveSearch = createAction('CLEAR_ACTIVE_SEARCH')
@@ -37,7 +34,7 @@ export function resetForm () {
       dispatch(settingQueryParam(otpState.user.defaults))
     } else {
       // Get user overrides and apply to default query
-      const userOverrides = getItem('defaultQuery', {})
+      const userOverrides = coreUtils.storage.getItem('defaultQuery', {})
       const defaultQuery = Object.assign(
         getDefaultQuery(otpState.config),
         userOverrides
@@ -71,7 +68,7 @@ export function parseUrlQueryString (params = getUrlParams()) {
     Object.keys(params).forEach(key => {
       if (!key.startsWith('ui_')) planParams[key] = params[key]
     })
-    const searchId = params.ui_activeSearch || randId()
+    const searchId = params.ui_activeSearch || coreUtils.storage.randId()
     // Convert strings to numbers/objects and dispatch
     dispatch(
       setQueryParam(
@@ -91,10 +88,11 @@ let lastDebouncePlanTimeMs
 export function formChanged (oldQuery, newQuery) {
   return function (dispatch, getState) {
     const otpState = getState().otp
+    const isMobile = coreUtils.ui.isMobile()
 
     // If departArrive is set to 'NOW', update the query time to current
     if (otpState.currentQuery && otpState.currentQuery.departArrive === 'NOW') {
-      dispatch(settingQueryParam({ time: moment().format(OTP_API_TIME_FORMAT) }))
+      dispatch(settingQueryParam({ time: moment().format(coreUtils.time.OTP_API_TIME_FORMAT) }))
     }
 
     // Determine if either from/to location has changed
@@ -113,7 +111,7 @@ export function formChanged (oldQuery, newQuery) {
     // either location changes only if not currently on welcome screen (otherwise
     // when the current position is auto-set the screen will change unexpectedly).
     if (
-      isMobile() &&
+      isMobile &&
       (fromChanged || toChanged) &&
       otpState.ui.mobileScreen !== MobileScreens.WELCOME_SCREEN
     ) {
@@ -125,8 +123,8 @@ export function formChanged (oldQuery, newQuery) {
     const { autoPlan, debouncePlanTimeMs } = otpState.config
     const updatePlan =
       autoPlan ||
-      (!isMobile() && oneLocationChanged) || // TODO: make autoplan configurable at the parameter level?
-      (isMobile() && fromChanged && toChanged)
+      (!isMobile && oneLocationChanged) || // TODO: make autoplan configurable at the parameter level?
+      (isMobile && fromChanged && toChanged)
     if (updatePlan && queryIsValid(otpState)) { // trip plan should be made
       // check if debouncing function needs to be (re)created
       if (!debouncedPlanTrip || lastDebouncePlanTimeMs !== debouncePlanTimeMs) {
