+++ lib/components/app/responsive-webapp.js
@@ -13,8 +13,7 @@ import { formChanged, parseUrlQueryString } from '../../actions/form'
 import { getCurrentPosition, receivedPositionResponse } from '../../actions/location'
 import { setLocationToCurrent } from '../../actions/map'
 import { handleBackButtonPress, matchContentToUrl } from '../../actions/ui'
-import { getTitle } from '../../util/ui'
-import { getActiveItinerary } from '../../util/state'
+import { getActiveItinerary, getTitle } from '../../util/state'
 
 const { isMobile } = coreUtils.ui
