+++ lib/components/form/default-search-form.js
@@ -6,12 +6,6 @@ import TabbedFormPanel from './tabbed-form-panel'
 import SwitchButton from './switch-button'
 import defaultIcons from '../icons'
 
-// Use the icons props to override icons for certain modes in the mode selector panel.
-// If no icon is provided for a specific mode, the default OTP-ui icon will be used.
-// const customIcons = {
-//  TRANSIT: <TriMetModeIcon label='rail' />
-// }
-
 export default class DefaultSearchForm extends Component {
   static propTypes = {
     icons: PropTypes.object,
@@ -19,7 +13,7 @@ export default class DefaultSearchForm extends Component {
   }
 
   static defaultProps = {
-    icons: defaultIcons, // customIcons,
+    icons: defaultIcons,
     showFrom: true,
     showTo: true
   }
