+++ lib/components/narrative/narrative-routing-results.js
@@ -25,14 +25,14 @@ class NarrativeRoutingResults extends Component {
   }
 
   render () {
-    const { customIcons, error, itineraryClass, itineraryFooter, pending, routingType, itineraries, mainPanelContent } = this.props
+    const { customIcons, error, itineraryClass, itineraryFooter, pending, itineraries, mainPanelContent } = this.props
     if (pending) return <Loading />
     if (error) return <ErrorMessage />
     if (mainPanelContent) return null
 
     return (
-      routingType === 'ITINERARY' &&
-        <TabbedItineraries itineraryClass={itineraryClass} itineraryFooter={itineraryFooter} itineraries={itineraries} customIcons={customIcons} />
+      // TODO: If multiple routing types exist, do the check here.
+      <TabbedItineraries itineraryClass={itineraryClass} itineraryFooter={itineraryFooter} itineraries={itineraries} customIcons={customIcons} />
     )
   }
 }
