+++ lib/util/index.js
@@ -1,9 +1,11 @@
+import coreUtils from '@opentripplanner/core-utils'
 import * as distance from './distance'
 import * as itinerary from './itinerary'
 import * as state from './state'
 import * as ui from './ui'
 
 const OtpUtils = {
+  ...coreUtils,
   distance,
   itinerary,
   state,
